backend/requirements.txt
# ========================================
# ChatBI Agent - 生产级依赖清单
# Python 版本: 3.13.0
# 最后更新: 2026-02-24
# ========================================

# === 核心 Web 框架 ===
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
pydantic-settings==2.12.0
starlette==0.38.6
aioodbc>=0.5.0
watchfiles>=0.24.0

# === LangChain 生态系统 ===
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.24
langgraph==0.2.76
langgraph-checkpoint==2.1.2
langgraph-checkpoint-postgres>=3.0.4
langsmith==0.1.143

# === LLM 提供商 ===
# Anthropic Claude
langchain-anthropic==0.2.4
anthropic==0.39.0

# OpenAI (可选)
langchain-openai==0.3.19
openai==1.109.1


# === 嵌入模型 ===
voyageai==0.2.4

# === 向量数据库 ===
langchain-pinecone==0.2.13
pinecone==7.3.0
pinecone-client==5.0.1

# === 数据库驱动 ===
sqlalchemy==2.0.36
greenlet==3.3.0  # SQLAlchemy 异步支持所需
aiomysql==0.2.0
aiosqlite==0.19.0
mysql-connector-python==9.1.0
PyMySQL==1.1.2
asyncpg==0.30.0  # PostgreSQL 异步支持
psycopg-binary==3.3.3 # PostgreSQL binary driver for development
psycopg-pool>=3.2.0   # PostgreSQL connection pool

# === 认证和安全 ===
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9
cryptography==42.0.5

# === 缓存 ===
redis==5.2.0
aioredis==2.0.1

# === HTTP 客户端 ===
httpx==0.28.1
httpx-sse==0.4.0
aiohttp==3.13.3
aiohttp-retry==2.9.1
requests==2.32.5

# === 日志和监控 ===
structlog==24.4.0
prometheus-client==0.21.0

# === 重试和错误处理 ===
tenacity==9.0.0

# === 数据处理 ===
numpy==2.4.1
tiktoken==0.12.0
langchain-text-splitters==0.3.8
sqlglot>=20.0.0

# === 工具库 ===
python-dotenv==1.0.1
email-validator>=2.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
orjson==3.11.5

# === 可选：重排序 ===
cohere==5.11.2

# === 开发工具 ===
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.1.0
black==24.10.0
ruff==0.7.4
mypy==1.13.0
types-requests==2.32.4.20260107
